+++ packages/plugin-openapi/src/Plugin.ts
@@ -1,0 +1,9 @@
+const PLUGIN_NAME = "openapi";
+
+class Plugin {
+  static get id() {
+    return PLUGIN_NAME;
+  }
+}
+
+export default Plugin;